Refactoring config and logger
