--- backend/api/services/logger.js
@@ -1,21 +1,0 @@
-var winston = require('winston');
-const defaultFormat = winston.format.printf(({ level, message, label, timestamp }) => {
-  return `${timestamp} ${level.toUpperCase()}: ${message}`;
-});
-
-const logger = winston.createLogger({
-  level: 'info',
-  format: winston.format.combine(winston.format.timestamp(), defaultFormat),
-  defaultMeta: {},
-  transports: [
-    //
-    // - Write to all logs with level `info` and below to `combined.log`
-    // - Write all logs error (and below) to `error.log`.
-    //
-    new winston.transports.File({ filename: 'appdata/logs/error.log', level: 'error' }),
-    new winston.transports.File({ filename: 'appdata/logs/combined.log' }),
-    new winston.transports.Console({level: !debugMode ? 'info' : 'debug', name: 'console'})
-  ]
-});
-
-exports.logger = logger;
